requirements.txt
streamlit
python-dotenv
langchain
faiss-cpu
PyPDF2
google-generativeai
